+++ galsim/fits.py
@@ -1017,21 +1017,30 @@ class FitsHeader(object):
 
         >>> pyf_header = fits_header.header
 
+    A FitsHeader may be constructed from a file name, an open PyFits (or astropy.io.fits) HDUList
+    object, or a PyFits (or astropy.io.fits) Header object.  It can also be constructed with
+    no parameters, in which case a blank Header will be constructed with no keywords yet if 
+    you want to add the keywords you want by hand.
+
+        >>> h1 = galsim.FitsHeader(file_name = file_name)
+        >>> h2 = galsim.FitsHeader(header = header)
+        >>> h3 = galsim.FitsHeader(hdu_list = hdu_list)
+        >>> h4 = galsim.FitsHeader()
+
+    For convenience, the first parameter may be unnamed as either a header or a file_name:
+
+        >>> h1 = galsim.FitsHeader(file_name)
+        >>> h2 = galsim.FitsHeader(header)
+
     Constructor parameters:
 
-    @param header       A pyfits Header object or in fact any dict-like object.  [One of 
-                        `file_name`, `hdu_list` or `header` is required.  The `header` parameter
-                        may be given without the keyword name.  Also, if `header` is a string, it
-                        is interpreted as a file name, so a file name may be passed as an arg like
-                        most other galsim.fits read functions.]
-    @param file_name    The name of the file to read in.  [One of `file_name`, `hdu_list` or 
-                        `header` is required.]
+    @param header       A pyfits Header object or in fact any dict-like object.  [default: None]
+    @param file_name    The name of the file to read in.  [default: None]
     @param dir          Optionally a directory name can be provided if `file_name` does not 
                         already include it. [default: None]
     @param hdu_list     Either a `pyfits.HDUList`, a `pyfits.PrimaryHDU`, or `pyfits.ImageHDU`.
                         In the former case, the `hdu` in the list will be selected.  In the latter
-                        two cases, the `hdu` parameter is ignored.  [One of `file_name`, `hdu_list`
-                        or `header is required.]
+                        two cases, the `hdu` parameter is ignored.  [default: None]
     @param hdu          The number of the HDU to return.  [default: None, which means to return
                         either the primary or first extension as appropriate for the given
                         compression.  (e.g. for rice, the first extension is the one you normally
@@ -1071,9 +1080,6 @@ class FitsHeader(object):
             raise TypeError("Cannot provide both hdu_list and header to FitsHeader")
         if file_name and hdu_list:
             raise TypeError("Cannot provide both file_name and hdu_list to FitsHeader")
-        # The line below allows 'header' to be set to {} (an empty dict).
-        if not (file_name or hdu_list) and (header is None):
-            raise TypeError("Must provide one of header, file_name or hdu_list to FitsHeader")
 
         # Interpret a string header as though it were passed as file_name.
         if isinstance(header, basestring):
@@ -1114,7 +1120,8 @@ class FitsHeader(object):
             # Otherwise, header may be any kind of dict-like object.
             # update() should handle anything that acts like a dict.
             self.header = pyfits.Header()
-            self.update(header)
+            if header is not None:
+                self.update(header)
 
     # The rest of the functions are typical non-mutating functions for a dict, for which we just
     # pass the request along to self.header.